+++ src/main/java/newmtsa/synthesis/gr1/OTFDirectedControledSyntesisGR1.java
@@ -496,12 +496,12 @@ public class OTFDirectedControledSyntesisGR1 implements OTFDirectedControlledSyn
         }
     }
 
-    public List<int[]> getFrontierWithFeatures() {
+    public List<float[]> getFrontierWithFeatures() {
         if (featureCompute == null)
             throw new IllegalStateException(
                     "No FeatureCompute registered — use the constructor overload that accepts FeatureCompute");
         prunePending();
-        List<int[]> result = new ArrayList<>(pending.size());
+        List<float[]> result = new ArrayList<>(pending.size());
         for (ExtendedTransition tr : pending) result.add(featureCompute.compute(tr));
         return result;
     }
